Edited some typo's
Hi, thank you for creating this repo for beginners! People like you make code a little less scary :)

I saw a couple spelling errors in the html, so I made a change or two, if you'd like to see them. I'm not sure if this is where one should submit this type of thing, (code newbie and all) so feel free to delete the PR if needed. Thanks again for helping build confidence in junior devs!